#add _CRT_SECURE_NO_WARNINGS if(MSVC)     add_definitions(-D_CRT_SECURE_NO_WARNINGS) endif()

+++ ppgl/mesh.cpp
@@ -373,7 +373,7 @@ extern "C" PPGL_EXPORT void  CGAL_3D_Intersection_Rays_Mesh_C1_Bool(Vector3d1 ps
 	inters = Vector1b2(ps.size(), Vector1b1());
 	for (int i = 0; i < ps.size(); i++)
 	{
-		Functs::OutputIterInfo("CGAL_3D_Intersection_Rays_Mesh_Bool", ps.size(), (int)i,(int)100);
+		Functs::OutputIterInfo("CGAL_3D_Intersection_Rays_Mesh_Bool", (int)ps.size(), (int)i,(int)100);
 		Point_3 p3=VectorPoint3d(ps[i]);
 		for (int j = 0; j < nes[i].size(); j++)
 		{
@@ -400,7 +400,7 @@ extern "C" PPGL_EXPORT void CGAL_3D_Intersection_Rays_Mesh_C2_Bool(Vector3d1 ps,
 	inters = Vector1b2(ps.size(), Vector1b1());
 	for (int i = 0; i < ps.size(); i++)
 	{
-		Functs::OutputIterInfo("CGAL_3D_Intersection_Rays_Mesh_Bool", ps.size(), i, 100);
+		Functs::OutputIterInfo("CGAL_3D_Intersection_Rays_Mesh_Bool", (int)ps.size(), (int)i, (int)100);
 		Point_3 p3 = VectorPoint3d(ps[i]);
 		for (int j = 0; j < ns.size(); j++)
 		{
